requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiohttp-retry==2.8.3
aiosignal==1.3.1
aiosqlite==0.20.0
annotated-types==0.7.0
anyio==4.6.2.post1
async-timeout==4.0.3
attrs==24.2.0
bcrypt==4.0.1
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
databases==0.9.0
distro==1.9.0
dnspython==2.7.0
ecdsa==0.19.0
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.115.2
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
jiter==0.6.1
multidict==6.1.0
openai==1.51.2
passlib==1.7.4
propcache==0.2.0
pyasn1==0.6.1
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.9.0
python-dotenv==1.0.1
python-jose==3.3.0
requests==2.32.3
rsa==4.9
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.39.2
tqdm==4.66.5
twilio==9.3.3
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.31.1
websockets==13.1
yarl==1.15.2
